requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
authlib==1.3.0
requests==2.31.0
pandas==2.1.4
openpyxl==3.1.2
pywebpush==2.0.0
cryptography==42.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
